// src/auditoria/enums/accion-auditoria.enum.ts
export enum AccionAuditoria {
  CREAR = 'CREAR',
  EDITAR = 'EDITAR',
  DESACTIVAR = 'DESACTIVAR',
  REACTIVAR = 'REACTIVAR',
  ELIMINAR = 'ELIMINAR',
  ANULAR = 'ANULAR',
  ABRIR = 'ABRIR',
  CERRAR = 'CERRAR',
  AJUSTAR = 'AJUSTAR',
  EMITIR = 'EMITIR',
  CANCELAR = 'CANCELAR',
  REGISTRAR = 'REGISTRAR',
  CAMBIAR_PERMISOS = 'CAMBIAR_PERMISOS',
}